Move lifecycle handlers to its own sub package.
This will allow injecting it into the runtimes.
